+++ src/main/java/org/ibp/api/rest/study/dataset/DatasetGeneratorInput.java
@@ -1,4 +1,4 @@
-package org.ibp.api.rest.dataset;
+package org.ibp.api.rest.study.dataset;
 
 /**
  * Created by clarysabel on 10/24/18.
